divided code in multiple files, fixed intel script, added unit test
