+++ test_files/intel_gpu_output.txt
@@ -1,0 +1,4 @@
+ Freq MHz      IRQ RC6     Power W           RCS/0           BCS/0           VCS/0          VECS/0 
+ req  act       /s   %   gpu   pkg       %  se  wa       %  se  wa       %  se  wa       %  se  wa 
+   0    0        0   0  0.00  0.00    0.00   0   0    0.00   0   0    0.00   0   0    0.00   0   0 
+   5    5       11  97  0.06  8.23    2.00   3   4    5.00   6   7    8.00   9   0    1.00   2   3
